+++ openff/pablo/_tests/unit/test_pdb_data.py
@@ -1214,55 +1214,3 @@ def test_can_parse_cif(pdbxfn: Path):
     assert data.line_no == list(range(2077, 18797))
 
     assert sum(1 for _ in data.residue_indices) == 34
-
-
-def test_get_residue_map_creates_complete_single_residue_graph(
-    cys_match: ResidueMatch,
-    cys_data: PdbData,
-):
-    graph, atoms, bonds = cys_data.get_residue_graph([[cys_match]])
-    assert len(atoms) == graph.n_nodes
-    assert len(bonds) == graph.n_edges
-    assert not any(atom is None for atom in atoms.values())
-    assert not any(bond is None for bond in bonds.values())
-    assert set(atoms.values()) == set(cys_match.residue_definition.atoms)
-    assert set(bonds.values()) == set(cys_match.residue_definition.bonds)
-
-
-def test_get_residue_map_mixed():
-    data = PdbData(conects=[{1}, {0}])
-    resdef = ResidueDefinition(
-        residue_name="AAA",
-        description="demo",
-        atoms=(AtomDefinition.with_defaults("A", "A"),),
-        bonds=(),
-        virtual_sites=(),
-        crosslink=None,
-        linking_bond=BondDefinition.with_defaults("A", "B"),
-    )
-    match = ResidueMatch(
-        residue_definition=resdef,
-        index_to_atomdef={0: resdef.atoms[0]},
-        vsite_idcs=(),
-        prior_bond_idcs=None,
-        posterior_bond_idcs=(0, 1),
-        crosslink_idcs=None,
-    )
-    mismatch = NoResidueDefinitions("BBB", {1: None})
-
-    graph, atoms, bonds = data.get_residue_graph([[match], [mismatch]])
-
-    assert graph.n_nodes == 2
-    assert graph.n_edges == 1
-    assert atoms == {0: resdef.atoms[0], 1: None}
-    assert bonds == {(0, 1): resdef.linking_bond}
-
-
-def test_get_residue_map_vicinal_disulfide(vicinal_disulfide_data: PdbData):
-    matches = vicinal_disulfide_data.match_residues(CCD_RESIDUE_DEFINITION_CACHE, ())
-    graph, atoms, bonds = vicinal_disulfide_data.get_residue_graph(matches)
-    assert len(atoms) == graph.n_nodes
-    assert len(bonds) == graph.n_edges
-    assert not any(atom is None for atom in atoms.values())
-    assert not any(bond is None for bond in bonds.values())
-    assert graph.is_connected()
